fixup: deduplicate cancellable-promise.ts between crud/aggregations
