+++ packages/compass-aggregations/package.json
@@ -41,6 +41,7 @@
     "@mongodb-js/compass-crud": "^13.3.0",
     "@mongodb-js/compass-editor": "^0.2.0",
     "@mongodb-js/compass-logging": "^1.1.1",
+    "@mongodb-js/compass-utils": "^0.1.1",
     "@mongodb-js/explain-plan-helper": "^1.0.1",
     "@mongodb-js/mongodb-constants": "^0.1.4",
     "@mongodb-js/mongodb-redux-common": "^2.0.2",
@@ -100,6 +101,7 @@
     "@mongodb-js/compass-crud": "^13.3.0",
     "@mongodb-js/compass-editor": "^0.2.0",
     "@mongodb-js/compass-logging": "^1.1.1",
+    "@mongodb-js/compass-utils": "^0.1.1",
     "@mongodb-js/explain-plan-helper": "^1.0.1",
     "@mongodb-js/mongodb-constants": "^0.1.4",
     "@mongodb-js/mongodb-redux-common": "^2.0.2",
