+++ packages/compass-aggregations/src/modules/aggregation.ts
@@ -3,7 +3,7 @@ import type { AggregateOptions, Document, MongoServerError } from 'mongodb';
 import type { PipelineBuilderThunkAction } from '.';
 import { DEFAULT_MAX_TIME_MS } from '../constants';
 import { globalAppRegistryEmit } from '@mongodb-js/mongodb-redux-common/app-registry';
-import { PROMISE_CANCELLED_ERROR } from '../utils/cancellable-promise';
+import { isCancelError } from '@mongodb-js/compass-utils'
 import { aggregatePipeline } from '../utils/cancellable-aggregation';
 import { ActionTypes as WorkspaceActionTypes } from './workspace';
 import { createLoggerAndTelemetry } from '@mongodb-js/compass-logging';
@@ -325,7 +325,7 @@ const fetchAggregationData = (
       }
     } catch (e) {
       // User cancel is handled in cancelAggregation
-      if ((e as Error).name === PROMISE_CANCELLED_ERROR) {
+      if (isCancelError(e)) {
         return;
       }
       // Server errors are surfaced to the user
