+++ packages/compass-crud/package.json
@@ -67,6 +67,7 @@
     "@mongodb-js/compass-components": "^1.3.0",
     "@mongodb-js/compass-editor": "^0.2.0",
     "@mongodb-js/compass-logging": "^1.1.1",
+    "@mongodb-js/compass-utils": "^0.1.1",
     "bson": "^4.4.1",
     "compass-preferences-model": "^2.3.0",
     "hadron-document": "^8.0.1",
@@ -113,6 +114,7 @@
     "@mongodb-js/compass-components": "^1.3.0",
     "@mongodb-js/compass-editor": "^0.2.0",
     "@mongodb-js/compass-logging": "^1.1.1",
+    "@mongodb-js/compass-utils": "^0.1.1",
     "bson": "^4.4.1",
     "compass-preferences-model": "^2.3.0",
     "hadron-document": "^8.0.1",
