+++ packages/compass-crud/src/utils/cancellable-queries.ts
@@ -2,14 +2,13 @@ import createLogger from '@mongodb-js/compass-logging';
 import { capMaxTimeMSAtPreferenceLimit } from 'compass-preferences-model';
 import type { DataService } from 'mongodb-data-service';
 import type { BSONObject } from '../stores/crud-store';
+import {
+  createCancelError,
+  isCancelError,
+  raceWithAbort,
+} from '@mongodb-js/compass-utils';
 
 const { log, mongoLogId, debug } = createLogger('cancellable-queries');
-
-/**
- * The error message to use whenever the user cancels the queries that are in
- * progress.
- */
-export const OPERATION_CANCELLED_MESSAGE = 'The operation was cancelled.';
 
 export async function findDocuments(
   dataService: DataService,
@@ -23,7 +22,7 @@ export async function findDocuments(
   } & Omit<Parameters<typeof dataService.fetch>[2], 'sort'>
 ): Promise<BSONObject[]> {
   if (signal.aborted) {
-    throw new Error(OPERATION_CANCELLED_MESSAGE);
+    throw createCancelError();
   }
 
   const cursor = dataService.fetch(ns, filter, options);
@@ -59,7 +58,7 @@ export async function countDocuments(
   >[2]
 ): Promise<number> {
   if (signal.aborted) {
-    throw new Error(OPERATION_CANCELLED_MESSAGE);
+    throw createCancelError();
   }
 
   const opts = {
@@ -101,7 +100,7 @@ export async function countDocuments(
     result = array.length ? array[0].count : 0;
   } catch (err: any) {
     // rethrow if we aborted along the way
-    if (err.message === OPERATION_CANCELLED_MESSAGE) {
+    if (isCancelError(err)) {
       throw err;
     }
 
@@ -130,7 +129,7 @@ export async function fetchShardingKeys(
 ): Promise<BSONObject> {
   // best practise is to first check if the signal wasn't already aborted
   if (signal.aborted) {
-    throw new Error(OPERATION_CANCELLED_MESSAGE);
+    throw createCancelError();
   }
 
   const cursor = dataService.fetch(
@@ -151,7 +150,7 @@ export async function fetchShardingKeys(
     configDocs = await raceWithAbort(cursor.toArray(), signal);
   } catch (err: any) {
     // rethrow if we aborted along the way
-    if (err.message === OPERATION_CANCELLED_MESSAGE) {
+    if (isCancelError(err)) {
       throw err;
     }
 
@@ -169,60 +168,4 @@ export async function fetchShardingKeys(
   signal.removeEventListener('abort', abort);
 
   return configDocs.length ? configDocs[0].key : {};
-}
-
-/*
- * Return a promise you can race (just like a timeout from timeouts/promises).
- * It will reject if abortSignal triggers before successSignal
- */
-function abortablePromise(
-  abortSignal: AbortSignal,
-  successSignal: AbortSignal
-): Promise<never> {
-  let reject: (err: Error) => void;
-
-  const promise = new Promise<never>(function (resolve, _reject) {
-    reject = _reject;
-  });
-
-  const abort = () => {
-    // if this task aborts it will never succeed, so clean up that event listener
-    // (abortSignal's event handler is already removed due to { once: true })
-    successSignal.removeEventListener('abort', succeed);
-
-    reject(new Error(OPERATION_CANCELLED_MESSAGE));
-  };
-
-  const succeed = () => {
-    // if this task succeeds it will never abort, so clean up that event listener
-    // (successSignal's event handler is already removed due to { once: true })
-    abortSignal.removeEventListener('abort', abort);
-  };
-
-  abortSignal.addEventListener('abort', abort, { once: true });
-  successSignal.addEventListener('abort', succeed, { once: true });
-
-  return promise;
-}
-
-/*
- * We need a promise that will reject as soon as the operation is aborted since
- * closing the cursor isn't enough to immediately make the cursor method's
- * promise reject.
- */
-async function raceWithAbort<T>(
-  promise: Promise<T>,
-  signal: AbortSignal
-): Promise<T> {
-  const successController = new AbortController();
-  const abortPromise = abortablePromise(signal, successController.signal);
-  try {
-    return await Promise.race([abortPromise, promise]);
-  } finally {
-    if (!signal.aborted) {
-      // either the operation succeeded or it failed because of some error
-      // that's not an abort
-      successController.abort();
-    }
-  }
 }
