+++ packages/compass-crud/src/utils/index.ts
@@ -2,7 +2,6 @@ export {
   findDocuments,
   countDocuments,
   fetchShardingKeys,
-  OPERATION_CANCELLED_MESSAGE,
 } from './cancellable-queries';
 
 /**
