+++ packages/compass-utils/src/index.ts
@@ -3,3 +3,9 @@ export {
   promisifyAmpersandMethod,
 } from './promisify-ampersand-method';
 export { StoragePaths, getStoragePaths } from './get-storage-paths';
+export {
+  raceWithAbort,
+  cancellableWait,
+  createCancelError,
+  isCancelError,
+} from './cancellable-promise';
